don't rerun hyperparameter search if model exists.

+++ src/hyperparameter_search.py
@@ -10,7 +10,8 @@ class HPModelBuilder:
 	def __init__(self, normalizer):
 		self.normalizer = normalizer
 	
-	def build_model_from_hyperparameters (self, hyperparameters: kt.engine.hyperparameters.HyperParameters) -> tf.keras.Sequential:
+	def build_model_from_hyperparameters (self, 
+		hyperparameters: kt.engine.hyperparameters.HyperParameters) -> tf.keras.Sequential:
 
 		# define solution space for fixed-length hyperparameters
 		hidden_layers = hyperparameters.get("hidden_layers")
@@ -28,27 +29,32 @@ def load_data(data_file: str) -> pd.DataFrame:
 	return pd.read_csv(data_file, index_col="Datetime", parse_dates=True)
 
 def generate_search_space(min_hidden_layers: int,
-						max_hidden_layers: int,
-						min_learning_rate: int,
-						max_learning_rate: int,
-						hidden_layer_size_options: int) -> kt.HyperParameters:
+	max_hidden_layers: int,
+	min_learning_rate: int,
+	max_learning_rate: int,
+	hidden_layer_size_options: int) -> kt.HyperParameters:
 	
-		hyperparameters = kt.HyperParameters()
+	hyperparameters = kt.HyperParameters()
 		
-		# define search space for fixed-length hyperparameters
-		hidden_layers = hyperparameters.Int("hidden_layers", min_hidden_layers, max_hidden_layers)
-		learning_rate = hyperparameters.Float("learning_rate", min_value=min_learning_rate, max_value=max_learning_rate, sampling="log")
+	# define search space for fixed-length hyperparameters
+	hidden_layers = hyperparameters.Int("hidden_layers", 
+		min_hidden_layers, 
+		max_hidden_layers)
+	learning_rate = hyperparameters.Float("learning_rate", 
+		min_value=min_learning_rate, 
+		max_value=max_learning_rate, 
+		sampling="log")
 		
-		for i in range(max_hidden_layers):
-			hyperparameters.Choice(f"units_{i}", hidden_layer_size_options)
-		
-		return hyperparameters
+	for i in range(max_hidden_layers):
+		hyperparameters.Choice(f"units_{i}", hidden_layer_size_options)
+	
+	return hyperparameters
 
 def get_tuner(model_builder: HPModelBuilder, 
-			hp_search_space: kt.HyperParameters,
-			search_dir: str, 
-			search_name: str, 
-			search_trials:int) -> kt.BayesianOptimization:
+	hp_search_space: kt.HyperParameters,
+	search_dir: str, 
+	search_name: str, 
+	search_trials:int) -> kt.BayesianOptimization:
 	'''
 	Create or load an instance of a keras hyperparameter tuning object.
 	For a given tuner, no overwrite will occur (even with changed parameters).
@@ -61,18 +67,17 @@ def get_tuner(model_builder: HPModelBuilder,
 		max_trials=search_trials,
 		directory=search_dir,
 		project_name=search_name,
-		hyperparameters=hp_search_space
-	)
+		hyperparameters=hp_search_space)
 
 	return tuner
 
 def search(tuner: kt.BayesianOptimization, 
-			train_features: pd.DataFrame, 
-			train_labels: pd.DataFrame, 
-			val_features: pd.DataFrame, 
-			val_labels: pd.DataFrame,
-			max_epochs: int,
-			early_stopping_patience: int):
+	train_features: pd.DataFrame, 
+	train_labels: pd.DataFrame, 
+	val_features: pd.DataFrame, 
+	val_labels: pd.DataFrame,
+	max_epochs: int,
+	early_stopping_patience: int):
 
 	# callback
 	early_stopping_callback = tf.keras.callbacks.EarlyStopping(monitor='val_loss', patience=early_stopping_patience)
@@ -80,11 +85,11 @@ def search(tuner: kt.BayesianOptimization,
 
 	# search
 	tuner.search(train_features, 
-			train_labels,
-			epochs=max_epochs,
-			validation_data=(val_features, val_labels),
-			verbose=False,
-			callbacks=[early_stopping_callback, tensorboard_callback])
+		train_labels,
+		epochs=max_epochs,
+		validation_data=(val_features, val_labels),
+		verbose=False,
+		callbacks=[early_stopping_callback, tensorboard_callback])
 
 	return None
 
@@ -121,10 +126,10 @@ def hyperparameter_search(config: dict) -> None:
 
 	# get hyperparameters
 	hp_search_space = generate_search_space(config["hp_min_hidden_layers"],
-											config["hp_max_hidden_layers"],
-											config["hp_min_learning_rate"],
-											config["hp_max_learning_rate"],
-											config["hp_hidden_layer_size_choices"])
+		config["hp_max_hidden_layers"],
+		config["hp_min_learning_rate"],
+		config["hp_max_learning_rate"],
+		config["hp_hidden_layer_size_choices"])
 
 	# get normalizer 
 	normalizer = get_normalization_layer(train_features)
@@ -132,19 +137,19 @@ def hyperparameter_search(config: dict) -> None:
 
 	# get tuner
 	tuner = get_tuner(model_builder,
-						hp_search_space,
-						config["hyperparameter_search_dir"],
-						config["hyperparameter_search_name"],
-						config["hp_search_trials"])
+		hp_search_space,
+		config["hyperparameter_search_dir"],
+		config["hyperparameter_search_name"],
+		config["hp_search_trials"])
 
 	# search
 	search(tuner, 
-			train_features, 
-			train_labels,
-			val_features,
-			val_labels,
-			config["ann_max_epochs"],
-			config["ann_early_stopping_patience"])
+		train_features, 
+		train_labels,
+		val_features,
+		val_labels,
+		config["ann_max_epochs"],
+		config["ann_early_stopping_patience"])
 
 	# get metadata from best model
 	best_hyperparameters = get_best_hyperparameters(tuner)
@@ -152,17 +157,18 @@ def hyperparameter_search(config: dict) -> None:
 	best_hyperparameters_series.to_csv(config["ann_summary_file"], header=False)
 	
 	# save model, history, and hyperparameters
-	model = model_builder.build_model_from_hyperparameters(best_hyperparameters)
-	history = train_model(model, 
-						train_features, 
-						train_labels, 
-						val_features, 
-						val_labels,
-						config["ann_max_epochs"],
-						config["ann_early_stopping_patience"])
+	if not os.path.exists(config["ann_model_file"])
+		model = model_builder.build_model_from_hyperparameters(best_hyperparameters)
+		history = train_model(model, 
+			train_features, 
+			train_labels, 
+			val_features, 
+			val_labels,
+			config["ann_max_epochs"],
+			config["ann_early_stopping_patience"])
 
-	model.save(config["ann_model_file"])
-	history_df = extract_history_to_dataframe(history)
-	history_df.to_csv(config["ann_history_file"])	
+		model.save(config["ann_model_file"])
+		history_df = extract_history_to_dataframe(history)
+		history_df.to_csv(config["ann_history_file"])	
 
 	return None
